Storage/VDI: Fix possible corruption when compacting an image with differencing images on a 32bit system caused by an integer overflow
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@40032 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Storage/VDI.cpp
@@ -2508,7 +2508,7 @@ static int vdiCompact(void *pBackendData, unsigned uPercentStart,
                 }
                 else if (pfnParentRead)
                 {
-                    rc = pfnParentRead(pvParent, i * cbBlock, pvBuf, cbBlock);
+                    rc = pfnParentRead(pvParent, (uint64_t)i * cbBlock, pvBuf, cbBlock);
                     if (RT_FAILURE(rc))
                         break;
                     if (!memcmp(pvTmp, pvBuf, cbBlock))
